+++ cherrypy/_cpreqbody.py
@@ -135,7 +135,7 @@ import six
 import cheroot.server
 
 import cherrypy
-from cherrypy._cpcompat import text_or_bytes, ntou, unquote_qs
+from cherrypy._cpcompat import text_or_bytes, ntou, unquote
 from cherrypy.lib import httputil
 
 
@@ -473,14 +473,7 @@ class Entity(object):
             if 'filename*' in disp.params:
                 # @see https://tools.ietf.org/html/rfc5987
                 encoding, lang, filename = disp.params['filename*'].split("'")
-                self.filename = unquote_qs(str(filename), encoding)
-
-    # The 'type' attribute is deprecated in 3.2; remove it in 3.3.
-    type = property(
-        lambda self: self.content_type,
-        doc='A deprecated alias for '
-            ':attr:`content_type<cherrypy._cpreqbody.Entity.content_type>`.'
-    )
+                self.filename = unquote(str(filename), encoding)
 
     def read(self, size=None, fp_out=None):
         return self.fp.read(size, fp_out)