+++ cherrypy/_cptools.py
@@ -22,12 +22,12 @@ Tools may be implemented as any object with a namespace. The builtins
 are generally either modules or instances of the tools.Tool class.
 """
 
-import sys
+import six
 
 import cherrypy
 from cherrypy._helper import expose
 
-from cherrypy.lib import cptools, encoding, auth, static, jsontools
+from cherrypy.lib import cptools, encoding, static, jsontools
 from cherrypy.lib import sessions as _sessions, xmlrpcutil as _xmlrpc
 from cherrypy.lib import caching as _caching
 from cherrypy.lib import auth_basic, auth_digest
@@ -37,7 +37,7 @@ def _getargs(func):
     """Return the names of all static arguments to the given function."""
     # Use this instead of importing inspect for less mem overhead.
     import types
-    if sys.version_info >= (3, 0):
+    if six.PY3:
         if isinstance(func, types.MethodType):
             func = func.__func__
         co = func.__code__
@@ -71,12 +71,13 @@ class Tool(object):
         self.__doc__ = self.callable.__doc__
         self._setargs()
 
-    def _get_on(self):
+    @property
+    def on(self):
         raise AttributeError(_attr_error)
 
-    def _set_on(self, value):
+    @on.setter
+    def on(self, value):
         raise AttributeError(_attr_error)
-    on = property(_get_on, _set_on)
 
     def _setargs(self):
         """Copy func parameter names to obj attributes."""
@@ -494,8 +495,6 @@ _d.caching = CachingTool('before_handler', _caching.get, 'caching')
 _d.expires = Tool('before_finalize', _caching.expires)
 _d.ignore_headers = Tool('before_request_body', cptools.ignore_headers)
 _d.referer = Tool('before_request_body', cptools.referer)
-_d.basic_auth = Tool('on_start_resource', auth.basic_auth)
-_d.digest_auth = Tool('on_start_resource', auth.digest_auth)
 _d.trailing_slash = Tool('before_handler', cptools.trailing_slash, priority=60)
 _d.flatten = Tool('before_finalize', cptools.flatten)
 _d.accept = Tool('on_start_resource', cptools.accept)
@@ -507,4 +506,4 @@ _d.auth_basic = Tool('before_handler', auth_basic.basic_auth, priority=1)
 _d.auth_digest = Tool('before_handler', auth_digest.digest_auth, priority=1)
 _d.params = Tool('before_handler', cptools.convert_params, priority=15)
 
-del _d, cptools, encoding, auth, static
+del _d, cptools, encoding, static